fix: update pdf preview module paths and selectors
This commit includes a refactoring of the PDF Preview module. The import paths and class type in
PdfPreviewBrick.js have been updated from 'namefruits/namereport-service' to 'quiqqer/bricks'.
Along with this, the selection for the PDF container was changed from 'pdf-preview' to
'pdf-preview-container'. An redundant line appending the 'pdfContainer' to the element has also
been removed. Additionally, the corresponding HTML and PHP files referencing these paths and
selectors have also been updated for consistency.